Pings now work locally, but still need to test over server
And also make them occur even when the receiving player doesn't have the mod installed. Perhaps I should pretend that the exit exists in a certain location and spawn something there? but then it also needs to be removed at a later time... Perhaps the removing and the creation of the exit can happen using RPC?