Mod Settings exists and logs!
It exists under the multiplayer button, and it logs the correct settings data (so far, not completely tested). In the near future, the button will be used to open a GUIPanel that will consist of all of the Mod settings. Currently, settings are properly ignored, and at least Enabled is displayed correctly when logging!