Began work on ModSettingsPopupPanel
Various formatting of buttons, simplified DustDevil.cs to instead simply call the Panel's show method. Constructing the Panel on the loading of the menu to also save performance, not to mention the buttons are in slightly awkward locations, and still need to learn how tabling works for placing the settings using a scroll. Also the panel is a much larger size than expected, as are all of the buttons.