Fixed naming, fields collection, improvements on menu
However, the toggles now don't appear at all, this is probably because of some strange AgeTransform parenting/childing that is happening. Names have been updated to include the mod name so that there are no duplicates, menu still appears large, but now it is a custom menu (as in, it isn't an OptionsPanel) but it still seems to contain some of the Options. Perhaps need to fiddle with how it is registered when it is displayed? To fix the attribute problem, we likely need to downgrade to target the right framework: .NET 2.something (idr if it is core or standard or framework either). Also need to get those buttons showing, wherever they may be.