Good progress, finally got the same size panel
But... The panel is in the incorrect position, the various buttons, pieces, etc. are improperly scaled (unless forced), buttons for the menu are different than for the OptionsPanel (wider/taller), all of the settings appear in one spot and do not move, even when their AgeTransforms are changed... Does this change persist across a refresh? are they forced by some unknown parent?