Need to determine what is going wrong with the settingspanel
Is it the frame? the parenting? the children components? the fact that there is no left side? the location of the object in its entirety? the offset of the panel?
And if it is one of those things, test changing them to see what happens, if anything. Maybe we can find a hackfix for placing the panel in the right place. But ofc the settings also have to be in the right place.