Updated DustDevilFramework to contain less debug, can load/unload!
Working and tested load/unload, but requires massive changes to ScadMod (all mods are impacted). Of course, the version for DustDevil needs to be updated to 4.0 soon, however there is a possible problem in that many-many-many TypeLoadExceptions occur with UnityEngine.GameObject.AddComponent (essentially every time it is called, a TypeLoadException happens)
This needs to be fixed before shipping because it is probably the source of most of my crashes so far.