Updated CustomModule it now works in game
BUT! It has yet to properly impersonate another module, so you can research it if you spot the cost difference, but no new icon appears. You can see it in the build menu, as it costs the proper amount, but no icon shows for the module. Clicking the module and attempting to place it causes a crash, as there seems to be no prefab loaded for the module either.
However, need to confirm more information on what is causing the CustomModule to crash.