So, seeded dungeons technically work
I want to move this stuff to DustDevil itself soon (especially SeedData and the ability to set seeds)
However, the way the seeding works is that every door you open and before every possible event, the exact same seed as the start are always applied. This means that you have the exact same seed for every door on a given floor, where that seed dictates the probability of a certain event. So, all the events are exactly the same, but I wish they still had higher variation, as currently, you can get a level with almost all of a certain thing depending on the seed you choose.