Removed Setting the same seed for each door, added it to TAS only
This is the result of a fruitful discussion with battledoog on Discord- We both believe that forcing a seed for each door (while possible) should not be part of the overall seed control for human players. In addition, added an option for TAS (unimplemented) that would force it to behave exactly like a human. This would mean that seeds would have to be a result of TAS actions (if desired at all). Also made it so only two important seeds are shown on the ship menu.