Added BepInEx functional DustDevil Framework!
Also includes simple logging with files, but there are some problems: Namely, it seems that there are sometimes two listeners and they aren't allowed to both write to the same object. Also, I would optimally have a way of formatting my log to look like BSIPA's, or even BepInEx's. But hey, SimpleCheats and the framework work with BepInEx!